### 2023-04-02 #5327 @Kevin-Mattheus-Moerman
@jackmwolf thanks for this submission. I am the AEiC on this track and here to help process the first steps. I have just triggered a scope review by the editorial board for this submission. This means I am asking for help to review if this work is deemed in scope for JOSS, and if it conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), in particular I feel I need some help checking if this qualifies as a "single functionality package".
The scope review should take about 2 weeks to complete.
