+++ README.md
@@ -76,6 +76,25 @@ vt_cate <- tunevt(
   data = tehtuner_example, Y = "Y", Trt = "Trt", step1 = "randomforest",
   step2 = "rtree", alpha0 = 0.2, p_reps = 100, ntree = 50
 )
+vt_cate
+#> Call:
+#> tunevt(data = tehtuner_example, Y = "Y", Trt = "Trt", step1 = "randomforest", 
+#>     step2 = "rtree", alpha0 = 0.2, p_reps = 100, ntree = 50)
+#> 
+#> Step 2 "rtree" model:
+#> n= 1000 
+#> 
+#> node), split, n, deviance, yval
+#>       * denotes terminal node
+#> 
+#> 1) root 1000 18372.4300  1.6830340  
+#>   2) V1< -3.008541 511  5643.7030 -0.3942647 *
+#>   3) V1>=-3.008541 489  8219.4140  3.8537890  
+#>     6) V3>=0.000282894 19   448.7299 -5.3806930 *
+#>     7) V3< 0.000282894 470  6084.9480  4.2270980 *
+#> 
+#> Approximate 0.8 quantile of the MNPP null distribution: 0.07695
+#> Observed MNPP: 0.2454,   p-value: < 2.22e-16
 ```
 
 The fitted Step 2 model can be accessed via `$vtmod`. In this case, as
@@ -129,10 +148,10 @@ the MNPP was above the 80th quantile.
 
 ### Running in Parallel
 
-Development version `0.1.1.9001` added the `parallel` option to
-`tunevt()` which allows the user to perform the permutation procedure in
-parallel to reduce computation times. Before doing so, you must register
-a parallel backend; see `?foreach::foreach` for more information.
+Version `0.2.0` added the `parallel` option to `tunevt()` which allows
+the user to perform the permutation procedure in parallel to reduce
+computation times. Before doing so, you must register a parallel
+backend; see `?foreach::foreach` for more information.
 
 For example, to carry out 100 permutations across 2 processors:
 
