### 2023-05-08 #5453 @editorialbot
[REVIEW]: tehtuner: An R package to fit and tune models for the conditional average treatment effect
**Submitting author:** @jackmwolf (<a href="http://orcid.org/0000-0002-8919-8740">Jack M Wolf</a>)
**Repository:** https://github.com/jackmwolf/tehtuner
**Branch with paper.md** (empty if default branch): 12-write-an-article-for-the-journal-of-open-source-software
**Version:** 0.3.1
