### 2023-05-23 jackmwolf/tehtuner#14 @elimillera
JOSS Review: @elimillera
Hey @jackmwolf, I've completed my review of the package. Overall a very good functioning and documented package. Below I've noted a couple small improvements.

- Automated tests
In the tests/testhat directory there is only a single test which is testing for the lack of an error. Consider making the tests more robust and adding tests for equality with `expect_equal` to verify the output is as expected and correct.

- Community guidelines
Consider adding a CONTRIBUTING.md to the root of the repository to clarify how people can contribute to the project. An example can be found [here](https://contributing.md/example/).

- State of the field
I've done a brief search for similar packages and I found [hdcate](https://cran.r-project.org/web/packages/hdcate/) which looks to have a similar purpose. Consider adding this to the state of the field section.

https://github.com/openjournals/joss-reviews/issues/5453
