### 2023-05-23 c2c16d4
Add citation for and discussion of hdcate R package to paper.md
