+++ paper.md
@@ -29,7 +29,7 @@ This is accomplished through a two-step procedure which first uses a flexible me
 
 # Statement of need
 
-Although there are several readily available R packages that can estimate the CATE [@hill_bayesian_2011; @vieille_avirtualtwins_2018; @tibshirani_grf_2022]; there are few, if any, at the time of writing which both estimate an interpretable model *and* guarantee controlled behavior when there is no treatment effect heterogeneity. While @vieille_avirtualtwins_2018 does provide an implementation of the original Virtual Twins manuscript, it does not support the methods in Steps 1 and 2 evaluated and recommended in @deng_practical_2023 such as super learner [@van_der_laan_super_2007] and is prone to overfitting when there is no effect heterogeneity.
+Although there are several readily available R packages that can estimate the CATE [@hill_bayesian_2011; @vieille_avirtualtwins_2018; @tibshirani_grf_2022]; there are few, if any, at the time of writing which both estimate an interpretable model *and* guarantee controlled behavior when there is no treatment effect heterogeneity. While @vieille_avirtualtwins_2018 does provide an implementation of the original Virtual Twins manuscript, it does not support the methods in Steps 1 and 2 evaluated and recommended in @deng_practical_2023 such as super learner [@van_der_laan_super_2007] and is prone to overfitting when there is no effect heterogeneity. Additionally, @fan_hdcate_2022 can estimate the effect of a given covariate on the treatment effect after marginalizing across all other covariates. This approach yields interpretable and well-behaved models but is unable to analyze the CATE in a multivariate manner.
 
 At this current time, `tehtuner` supports linear models fit via the LASSO [@tibshirani_regression_1996], MARS [@friedman_multivariate_1991], random forests [@breiman_random_2001], and super learner [@van_der_laan_super_2007] in Step 1 and linear  models tuned via the LASSO [@tibshirani_regression_1996], regression and classification trees [@breiman_classification_2017], and conditional inference trees [@hothorn_unbiased_2006] in Step 2. Comparative evaluations of these methods can be found in @wolf_permutation_2022 and @deng_practical_2023.
 
