+++ inst/REFERENCES.bib
@@ -171,3 +171,13 @@ Section: Statistical Applications in Genetics and Molecular Biology},
 	year = {2018},
 	annote = {R package version 1.0.1},
 }
+
+
+@book{fan_hdcate_2022,
+	title = {hdcate: {Estimation} of {Conditional} {Average} {Treatment} {Effects} with {High}-{Dimensional} {Data}},
+	url = {https://CRAN.R-project.org/package=hdcate},
+	author = {Fan, Qingliang and Hong, Hengzhao},
+	year = {2022},
+	annote = {R package version 0.1.0},
+}
+
