### 2023-05-26 9f48145
Address JOSS comments from #14 - Add automated tests for Step 1 wrapper functions and get_mnpp() - Create CONTRIBUTING.md
