+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: tehtuner
 Title: Fit and Tune Models to Detect Treatment Effect Heterogeneity
-Version: 0.3.0.9000
+Version: 0.3.0.9001
 Authors@R: 
     person(given = "Jack",
            family = "Wolf",
