### 2023-05-26 e71c2b7
Revise discussion of hdcate to mention confirmatory/exploratory analysis
