### 2023-05-26 jackmwolf/tehtuner#14 @jackmwolf
Hi @elimillera! Thank you for the suggestions- I have worked to address them all:

1. I have added automated tests for all of the internal Step 1 wrapper functions (e.g., `vt1_lasso`) to verify their predictions agree with Step 1 models manually fit by hand. I have also added a test to confirm that the proposed minimum null penalty parameter returned from `get_mnpp` results in a null Step 2 model for all possible Step 2 models. These tests live in [`test-main.R`](https://github.com/jackmwolf/tehtuner/blob/9f481459721ab8d71f4bbdc8bc63dc447ae43dfa/tests/testthat/test-main.R) (commit 9f48145).
2. I have modified the example you provided for this project and placed it in the root [`CONTRIBUTING.md`](https://github.com/jackmwolf/tehtuner/blob/9f481459721ab8d71f4bbdc8bc63dc447ae43dfa/CONTRIBUTING.md) (commit 9f48145).
3. Thank you for bringing `hdcate` to my attention. I have cited it in the state of the field as an alternative method for analyzing the conditional average treatment effect and noted how it is different from the Virtual Twins approach implemented in `tehtuner` (commits c2c16d4 and e71c2b7 in [branch 12-write-an-article-for-the-journal-of-open-source-software](https://github.com/jackmwolf/tehtuner/tree/12-write-an-article-for-the-journal-of-open-source-software)). I have verified that the paper renders a PDF without errors after these changes.

Please let me know if you have any further questions. Again, thank you for the feedback and help to improve the package.
