### 2023-06-07 #5453 @chartgerink
@jackmwolf - we got two reviews and they signed off on all the items on the checklist. I am happy to move this forward to the final stages. Would you mind taking a look at the author tasks in the linked comment? https://github.com/openjournals/joss-reviews/issues/5453#issuecomment-1580104816
