### 2023-06-08 c4669ed
Increase version to 0.3.1 corresponding to JOSS paper.
