### 2023-06-08 #5453 @jackmwolf
@chartgerink I have checked off the items assigned to me. The software release version is 0.3.1, the DOI is 10.5281/zenodo.8018280. Please let me know if there's anything else I can do to help. Thanks!

- [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **Version 0.3.1**
- [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. **DOI 10.5281/zenodo.8018280**
